Fix file per class
