+++ src/tic-tac-toe/Board.tsx
@@ -1,12 +1,7 @@
 import * as React from "react";
+import { Square } from "./Square";
 
-class Square extends React.Component<{ value: number }> {
-  render() {
-    return <button className="square">{this.props.value}</button>;
-  }
-}
-
-class Board extends React.Component {
+export class Board extends React.Component {
   static renderSquare(i: number) {
     return <Square value={i} />;
   }
@@ -31,22 +26,6 @@ class Board extends React.Component {
           {Board.renderSquare(6)}
           {Board.renderSquare(7)}
           {Board.renderSquare(8)}
-        </div>
-      </div>
-    );
-  }
-}
-
-export class Game extends React.Component {
-  render() {
-    return (
-      <div className="game">
-        <div className="game-board">
-          <Board />
-        </div>
-        <div className="game-info">
-          <div>{/* status */}</div>
-          <ol>{/* TODO */}</ol>
         </div>
       </div>
     );
