+++ src/tic-tac-toe/Square.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+export class Square extends React.Component<{ value: number }> {
+  render() {
+    return <button className="square">{this.props.value}</button>;
+  }
+}
